Add definition for custom scalar BigDecimal
